# backend/api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Obscurenv API
  description: |
    Zero-knowledge encrypted `.env` storage. The backend stores only opaque
    encrypted payloads and metadata. Never send plaintext env values,
    passphrases, derived keys, or raw encryption keys to this API.

    The CLI performs encryption/decryption locally. API calls should send only
    encrypted payloads. See `docs/api.md` for the developer guide.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Default local backend (override with OBE_API_URL / ADDR)
security: []
...
paths:
...
  /api/v1/user/profile:
    get:
      tags: [User]
      security: [{ bearerAuth: [] }]
      summary: Get the current user's profile
      operationId: getProfile
      responses:
        '200':
          description: Profile of the authenticated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
        '404':
          description: User not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    patch:
      tags: [User]
      security: [{ bearerAuth: [] }]
      summary: Update the current user's username
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username: { type: string, example: alice-dev }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404':
          description: User not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Username already taken
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/user/password:
    post:
      tags: [User]
      security: [{ bearerAuth: [] }]
      summary: Change the current user's password
      description: |
        Verifies the current password before replacing it. Existing API tokens
        remain valid after a password change.
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        '200':
          description: Password updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: password updated }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404':
          description: User not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/user:
    delete:
      tags: [User]
      security: [{ bearerAuth: [] }]
      summary: Delete the current user's account
      description: |
        Permanently deletes the account, its projects, environments, tokens,
        and activity logs. Requires an explicit `confirm: true`.
      operationId: deleteAccount
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeleteAccountRequest' }
      responses:
        '200':
          description: Account deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: account deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404':
          description: User not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API token returned by POST /auth/login or passkey login, e.g. `obe_tok_...`.
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, example: project not found }
...
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password: { type: string, format: password }
        new_password: { type: string, format: password, minLength: 8 }
    DeleteAccountRequest:
      type: object
      required: [confirm]
      properties:
        confirm: { type: boolean, description: Must be true to delete the account }
    UserProfile:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        username:
          type: [string, 'null']
        created_at: { type: string, format: date-time }
...